v0.3: interactive multi-turn agent loop
- Event schema v2: f64 confidence, per-run monotonic seq assigned by
  the session store, new event kinds (tool_call_requested,
  approval_requested/decided, turn_completed, input_received),
  tolerant reading of v1 logs, prompt recorded in run_started
- Tool protocol: OpenAI-compatible ToolDefinition/ToolCall/ToolResult;
  providers without tools capability reject tool requests with typed
  errors; OpenAI-compatible provider gains tools wire format
- ScriptedMockModel: JSON-scripted reply queue incl. tool calls,
  configurable via mock_script / FORGE_MOCK_SCRIPT
- ExecutionProvider gains file ops (read/write/edit/delete) with risk
  classification (read=safe, in-project write/edit=risky, delete or
  root-escape=destructive) through the same approval gating, plus
  approved-bypass methods invoked only after recorded approval
- AgentService multi-turn loop: graph+skill seeded context, tool
  dispatch, max_turns budget (config + --max-turns), in-loop approval
  pause/decide via per-run input channel, mid-loop cancellation
  (in-process token + cross-process .forge/runs/<id>.cancel marker),
  resume continuing completed runs in the same session
- Server: POST /v1/runs/:id/input delivers to paused runs (202/404/
  409), cancel aborts mid-loop, waiting_for_approval status, bounded
  run registry (1024, terminal-first eviction, store fallback)
- BDD: 6 new features (agent loop, approval, resume, server input,
  cancellation, event schema v1 compatibility); 16 features / 24
  scenarios / 93 steps, all offline
- README updated for v0.3; removed accidentally added prompt files

+++ crates/forge-core/src/router.rs
@@ -48,7 +48,7 @@ impl RoutingRequest {
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct RoutingDecision {
     pub selected_model: String,
-    pub confidence: f32,
+    pub confidence: f64,
     pub router_name: String,
     pub fallback_used: bool,
     pub reason: String,
